Confirmation objects use default instead of auto_now
auto_now fields cannot be overridden, which while it may work in all practical use-cases of these models, I would rather have more control.

Also abstracted the fetching of the Prosody domain as the default on the Prosody.host field; this avoids battling it out between dev and prod settings in the migration files.