Break email threading for notifications
By setting the X-Entity-Ref-ID header, we now break out of "conversation threading" (at least in Gmail) so that each message appears separately in the inbox. Additionally, we now vary the content of each message, so even if a client does thread the messages the user will still be able to see the full content of each message in the thread.