Rearrange template inheritance
We add another level to the template inheritance so we can quickly and easily define header and title content with one block in child templates. We have also tweaked the global base template to allow the header to be removed/replaced by child templates (which we are doing with the forum base templates).