Use an annotation to save us several DB hits
A simple annotation is all it takes to put the author of a post/thread on the Post object itself, rather than running back to the database for each one to fetch the User object. Note that we could do the same via select_related() instead, but this trims the data we fetch down a bit -- although we may have to backtrack later if we want to provide a user page linked from each post/topic!